Non adaptive k + engineered bath
Modifications:
chaged optimal k such that in increases as long as the measurement time isn't longer than 1/FWHM (using it as a lowerbound since experimentally we wouldn't have access to the quantum distribution and hence the actual value of T2*). Added peak detection so that if multiple peaks are detected, k is reduced.

Added 'engineered' bath based on proximity to a certain number of spins, see note in bath_tracking.py.

Still need to:
-time peakdetect.py
-clean the new bit added to spin bath script