Changed updaye values -D
Changed the p_k updates. adding a pi (independent of the ramsey result) matches the quantum distribution without the 'ad hoc' value of pi I had to add to the phase.

Regarding the prior, I've played around with different combinations of distributions, but if we start from a narrower distribution, it takes a few steps for the quantum distribution to catch up. The gaussian we start with right now fits the distribution well, since the quantum one is basically a 'partial' Gaussian after the first emeasurements. I'll keep playing around to see if there's a better fit.